:sparkles: feat: Implement global error handling
refs #34
